Move ReactComponent to Vue injector
